requirements.txt
streamlit
deepseek
faiss-cpu
pandas
numpy
scikit-learn
tqdm
seaborn
matplotlib
sentence-transformers
jupyter
